refactored ai part
