+++ src/commands/command_dispatcher.py
@@ -1,11 +1,12 @@
 from typing import Dict
 
+from ai.ai_commands import RunAIAssistantCommand
 from commands.command import Command, HelloCommand
 
 from notes_book.notes_commands import AddNoteCommand, SearchNoteByTilte, SearchNoteByTagCommand, EditNoteCommand, DeleteNoteCommand
 from address_book.address_book_commands import (AddBirthdayCommand, AddContactCommand, AllContactsCommand, ChangePhoneCommand,
                               ContactPhoneCommand, ShowBirthdayCommand, ShowBirthdaysCommand,
-                              RemoveContactCommand, EditContactCommand, RunAIAssistantCommand, SearchContactsCommand)
+                              RemoveContactCommand, EditContactCommand, SearchContactsCommand)
 from localization import get_text
