+++ src/notes_book/notes_commands.py
@@ -60,7 +60,7 @@ class SearchNoteByTagCommand(Command):
         for note in result:
             print(note.title)
 
-# class EditNoteCommand(Command):
+class EditNoteCommand(Command):
 
     def execute(self, **kwargs):
         notes = kwargs.get('notes')
@@ -91,44 +91,3 @@ class DeleteNoteCommand(Command):
             print(get_text("NOTE_NOT_FOUND")) 
         except ExitFromUserPrompt:
             print(get_text("NOTE_IS_NOT_DELETED"))
-
-
-# class SearchNoteByTitleCommand(Command):
-#     def execute(self, *args, **kwargs):
-#         query = input("Enter the query to search by title: ")
-#         results = kwargs['notes'].search_by_title(query)
-#         if results:
-#             print(get_text("SEARCH_RESULTS"))
-#             for result in results:
-#                 print(result)
-#         else:
-#             print(get_text("NO_NOTES_FOUND"))
-
-
-# class AddTagCommand(Command):
-
-#     def execute(self, *args, **kwargs):
-#         tags_input = input("Enter the tags separated by commas: ")
-#         tags = [tag.strip() for tag in tags_input.split(",")]
-#         for tag in tags:
-#             kwargs['tag_manager'].add_tag(tag)
-#         print(get_text("TAGS_ADDED"))
-        
-# class EditTagCommand(Command):
-
-#     def execute(self, *args, **kwargs):
-#         old_tag = input("Enter the old tag: ")
-#         new_tag = input("Enter the new tag: ")
-#         kwargs['tags'].edit_tags(Note_Tag(old_tag), Note_Tag(new_tag))
-
-# class SearchNoteByTagCommand(Command):
-
-#     def execute(self, *args, **kwargs):
-#         tag = input("Enter the tag to search by: ")
-#         results = kwargs['notes'].search_by_tag(tag)
-#         if results:
-#             print(get_text("SEARCH_RESULTS"))
-#             for result in results:
-#                 print(result)
-#         else:
-#             print(get_text("NO_NOTES_FOUND_BY_TAG"))
